Made the backpropagation compute one set of data at a time successfully
I found some problems with the prior code, basically there was accumulating error from not re intitializing the yarr array before using the feedforward function, also the error was being incorrectly computed as the output was not taken away from the desired output. Now training on the same data the error always reduces which is good, however the error does not quickly converge which is not a problem, training on 80 data sets to produce a value of 1 the output after training is 0.93 when the desired is one. Perhaps implementing a second order backpropagation will improve the convergence significantly, I need to also install file input of weights, inputs and actualoutputs.